Made Cognito's initiateAuth response type reflect API spec
The existing type says that all of the keys are optional, which isn't correct and requires additional type checking from the user.  What it should say is that it will either have an AuthenticationResult or it will have Session, ChallengeName, and ChallengeParameters.  This new type reflects that fact, so that if an if-statement checks that a `InitiateAuthResponse` object does not have AuthenticationResult, the compiler will know that it has all three other values.